Refactor plan builder into SRP modules
